Sink, Stream, Variable, Constant
